requirements.txt
numpy>=1.24.0
scipy>=1.11.0
trimesh>=4.0.0
rtree>=1.0.0
Pillow>=10.0.0
litemapy>=0.10.0b0
nbtlib>=2.0.4
gradio_client>=0.10.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.9
